feat: Allow for numbers in partial names
This gives us a way to reorder patterns. This removes the number from
the partial that is registered with Handlebars, as well as any numbers
you pass into a pattern helper.

I could see this creating problems if you have two patterns of the same
type with the same name. But if you run into this, you probably have other
organizational issues that should be sorted out.